faas/image-text-extractor/requirements.txt
pytesseract 
validators 
pillow